Refactor cyral_repository_conf_auth
