+++ cyral/internal/repository/confauth/constants.go
@@ -1,0 +1,11 @@
+package confauth
+
+const (
+	resourceName = "cyral_repository_conf_auth"
+
+	DefaultClientTLS    = "disable"
+	DefaultRepoTLS      = "disable"
+	AccessTokenAuthType = "ACCESS_TOKEN"
+	AwsIAMAuthType      = "AWS_IAM"
+	DefaultAuthType     = AccessTokenAuthType
+)
