+++ cyral/internal/repository/confauth/model.go
@@ -1,0 +1,92 @@
+package confauth
+
+import (
+	"errors"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
+
+var authTypes = []string{
+	AccessTokenAuthType,
+	AwsIAMAuthType,
+}
+
+type RepositoryConfAuthData struct {
+	RepoID           *string `json:"-"`
+	AllowNativeAuth  bool    `json:"allowNativeAuth"`
+	ClientTLS        string  `json:"clientTLS"`
+	IdentityProvider string  `json:"identityProvider"`
+	RepoTLS          string  `json:"repoTLS"`
+	AuthType         string  `json:"authType"`
+}
+
+func (data RepositoryConfAuthData) WriteToSchema(d *schema.ResourceData) error {
+	if data.RepoID != nil {
+		d.Set("repository_id", data.RepoID)
+	}
+
+	d.Set("allow_native_auth", data.AllowNativeAuth)
+
+	if err := data.isClientTLSValid(); err != nil {
+		panic(err)
+	}
+
+	d.Set("client_tls", data.ClientTLS)
+
+	d.Set("identity_provider", data.IdentityProvider)
+
+	if err := data.isRepoTLSValid(); err != nil {
+		panic(err)
+	}
+
+	d.Set("repo_tls", data.RepoTLS)
+
+	d.Set("auth_type", data.AuthType)
+
+	return nil
+}
+
+func (data *RepositoryConfAuthData) ReadFromSchema(d *schema.ResourceData) error {
+	if repoIdData, hasRepoId := d.GetOk("repository_id"); hasRepoId {
+		repoId := repoIdData.(string)
+		data.RepoID = &repoId
+	}
+
+	data.AllowNativeAuth = d.Get("allow_native_auth").(bool)
+	data.AuthType = d.Get("auth_type").(string)
+	data.ClientTLS = d.Get("client_tls").(string)
+	data.IdentityProvider = d.Get("identity_provider").(string)
+	data.RepoTLS = d.Get("repo_tls").(string)
+
+	return nil
+}
+
+func (data RepositoryConfAuthData) isClientTLSValid() error {
+	if !(data.ClientTLS == "enable" || data.ClientTLS == "disable" || data.ClientTLS == "enabledAndVerifyCertificate") {
+		return errors.New("invalid option to client_tls")
+	}
+	return nil
+}
+
+func (data RepositoryConfAuthData) isRepoTLSValid() error {
+	if !(data.RepoTLS == "enable" || data.RepoTLS == "disable" || data.RepoTLS == "enabledAndVerifyCertificate") {
+		return errors.New("invalid option to repo_tls")
+	}
+	return nil
+}
+
+type CreateRepositoryConfAuthResponse struct{}
+
+func (data CreateRepositoryConfAuthResponse) WriteToSchema(d *schema.ResourceData) error {
+	d.SetId(d.Get("repository_id").(string))
+	return nil
+}
+
+type ReadRepositoryConfAuthResponse struct {
+	AuthInfo RepositoryConfAuthData `json:"authInfo"`
+}
+
+func (data ReadRepositoryConfAuthResponse) WriteToSchema(d *schema.ResourceData) error {
+	data.AuthInfo.WriteToSchema(d)
+	return nil
+}
